+++ src/App.tsx
@@ -1,4 +1,9 @@
 import { useWindowDimensions } from "@fi-sci/misc";
+import {
+  AIContextProvider,
+  ComponentRegistrationForAI,
+  useAIContext,
+} from "./AIContext";
 import { useEffect } from "react";
 import SettingsIcon from "@mui/icons-material/Settings";
 import {
@@ -172,6 +177,45 @@ const LegacyUrlHandler = () => {
 const AppContent = () => {
   const navigate = useNavigate();
   const [searchParams] = useSearchParams();
+  const location = useLocation();
+  const { registerComponentForAI, unregisterComponentForAI } = useAIContext();
+
+  useEffect(() => {
+    const registration: ComponentRegistrationForAI = {
+      id: "App",
+      context: {
+        currentRoute: location.pathname,
+        params: Object.fromEntries(searchParams.entries()),
+      },
+      callbacks: [
+        {
+          id: "navigate",
+          description: aiRouteContextDescription,
+          parameters: {
+            route: {
+              type: "string",
+              description: "The route to navigate to. Be sure to provide this.",
+            },
+          },
+          callback: (parameters) => {
+            if (!parameters.route) {
+              console.warn("No route provided to navigate to", parameters);
+              return;
+            }
+            navigate(parameters.route);
+          },
+        },
+      ],
+    };
+    registerComponentForAI(registration);
+    return () => unregisterComponentForAI("route");
+  }, [
+    location,
+    searchParams,
+    navigate,
+    registerComponentForAI,
+    unregisterComponentForAI,
+  ]);
   const { width, height } = useWindowDimensions();
   const hideAppBar = searchParams.get("embedded") === "1";
   const appBarHeight = hideAppBar ? 0 : 50; // hard-coded to match the height of the AppBar
@@ -349,11 +393,25 @@ function App() {
   return (
     <ThemeProvider theme={theme}>
       <Router>
-        <LegacyUrlHandler />
-        <AppContent />
+        <AIContextProvider>
+          <LegacyUrlHandler />
+          <AppContent />
+        </AIContextProvider>
       </Router>
     </ThemeProvider>
   );
 }
+
+const aiRouteContextDescription = `
+Navigate to a new route in the Neurosift application.
+
+Possible routes are:
+/dandi (Browse DANDI)
+/dandiset/:dandisetId (View a DANDI dataset)
+/openneuro (Browse OpenNeuro)
+/openneuro-dataset/:datasetId (View an OpenNeuro dataset)
+/settings (View the settings page to set API keys)
+/guide (View the NWB Viewer Guide)
+`;
 
 export default App;